pyproject.toml
[project]
name = "bioteam-ai"
version = "0.1.0"
description = "Personal AI Science Team for Biology Research"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "JangKeun Kim" },
]

dependencies = [
    "anthropic>=0.79.0",
    "instructor>=1.14.0",
    "fastapi[standard]>=0.129.0",
    "uvicorn[standard]>=0.41.0",
    "sse-starlette>=3.2.0",
    "chromadb>=1.5.0",
    "langfuse>=3.14.0",
    "biopython>=1.86",
    "semanticscholar>=0.11.0",
    "alembic>=1.18.0",
    "sqlmodel>=0.0.34",
    "httpx>=0.28.0",
    "aiohttp>=3.13.0",
    "python-dotenv>=1.2.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.13.0",
    "PyYAML>=6.0.0",
    "arxiv>=2.1.0",
    "aiosmtplib>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]

[tool.pytest-asyncio]
mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501", "N806", "E402"]
